Feature: Add user shell prompt and read input
prompt prints a POS1 charster to ask for user input.
getline() is used to capture input from the standard input and write to the 'args' buffer.
The user input is then printed back to the user along with the number of bytes read.
Function definitions and library imports are in the shell.h header file